The task is to classify the article in two groups politicsNews and worldNews.
PART 1 - CountVectorizer, TF-IDF & Log Reg
Conclusion: First try was to apply Bag of words with default settings as well as Logistic Regression with default settings.
The cross validation accuracy was 0.935, not bad. Using GridSearchCV and finding best params was possible to push
the cross validation accuracy to 0.942. Testing this best model on the test set got accuracy of 0.936, not bad.

Trying to get more using Stop-words in Countvectorizer got the cross validation accuracy of 0.942 same as without Stop words.
Last try for rescaling features with TF-IDF with different n_grams-values got the cross validation of 0.946,
kind of the same as other methods.

PART 2 - Latent Dirichlet Allocation (LDA)
Conclusion: This approach was quite successful; it seems that the algorithm recognized two groups politics news amd world news.
We can see it in the print of some topic words.

PRINT OF ARTICLES NOT POSSIBLE?????